smoll-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.2.0
transformers==4.40.0
accelerate==0.25.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
bitsandbytes==0.41.3  # Untuk quantisasi (optional)